[PRMT-4524] Refactored exceptions
